Окончательные изменения

+++ backend/backend/api/views.py
@@ -1,26 +1,14 @@
-import time
-
-from PyPDF2 import PdfFileWriter, PdfWriter
-from django.contrib.auth import authenticate
-from fpdf import FPDF, HTMLMixin
-from reportlab.pdfgen.canvas import Canvas
+from fpdf import FPDF
 from rest_framework import status
-from rest_framework.generics import CreateAPIView, ListAPIView, GenericAPIView, RetrieveAPIView
+from rest_framework.generics import ListAPIView, RetrieveAPIView
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
-from rest_framework.permissions import IsAdminUser, IsAuthenticated
-import html
 from rest_framework.response import Response
-from rest_framework.authtoken.models import Token
 from api.models import CustomUser, SubscribingAuthors, Tag, Ingredient, IngredientForRecipe, Recipe, FavoritesList, \
     ShoppingList
 from api.pagination import PaginatorDefault
-from api.serializer import TagSerializers, IngredientSerializers, IngredientForRecipeSerializers, \
-    IngredientCreateSerializers, RecipeCreateSerializers, RecipeForShoppingList
+from api.serializer import TagSerializers, IngredientSerializers, RecipeCreateSerializers, RecipeForShoppingList
 from django.shortcuts import get_object_or_404
-import PyPDF2
-from datetime import datetime
-import pdfkit
-from reportlab.lib.pagesizes import A4
 
 
 class TagList(ListAPIView):
@@ -99,23 +87,49 @@ def get_lists(request):
     return subscribing_authors, favorites_list, shopping_list
 
 
+def check_filter(filter_is_favorite, is_in_shopping_cart, filter_author, filter_tags, recipe, shopping_list,
+                 favorites_list, author, tags):
+    if filter_is_favorite is not None and (
+            filter_is_favorite
+            and recipe.get('id') not in favorites_list
+            or not filter_is_favorite
+            and recipe.get('id') in favorites_list
+    ):
+        return True
+    if is_in_shopping_cart is not None and (
+            is_in_shopping_cart
+            and recipe.get('id') not in shopping_list
+            or not is_in_shopping_cart
+            and recipe.get('id') in shopping_list
+    ):
+        return True
+    if filter_author is not None and filter_author != author.id:
+        return True
+    if filter_tags is not None and filter_tags is not tags:
+        return True
+    return False
+
+
+def filters(request):
+    filter_is_favorite = int(request.GET.get('is_favorited', 0)) == 1 if request.GET.get(
+        'is_favorited', None) is not None else None
+    is_in_shopping_cart = None
+    if request.GET.get('is_in_shopping_cart', None) is not None:
+        is_in_shopping_cart = int(request.GET.get('is_in_shopping_cart', 0)) == 1
+    filter_author = int(request.GET.get('author')) if request.GET.get('author', None) is not None else None
+    filter_tags = int(request.GET.get('tags')) if request.GET.get('tags', None) is not None else None
+    return filter_is_favorite, is_in_shopping_cart, filter_author, filter_tags
+
+
 class CreateRecipe(ListAPIView, APIView):
     queryset = Recipe.objects.all()
     serializer_class = RecipeCreateSerializers
 
-    def filters(self, request):
-        filter_is_favorite = int(request.GET.get('is_favorited', 0)) == 1 if request.GET.get(
-            'is_favorited', None) is not None else None
-        is_in_shopping_cart = None
-        if request.GET.get('is_in_shopping_cart', None) is not None:
-            is_in_shopping_cart = int(request.GET.get('is_in_shopping_cart', 0)) == 1
-        filter_author = int(request.GET.get('author')) if request.GET.get('author', None) is not None else None
-        filter_tags = int(request.GET.get('tags')) if request.GET.get('tags', None) is not None else None
-        return filter_is_favorite, is_in_shopping_cart, filter_author, filter_tags
-
     def post(self, request):
+        if not request.user.is_authenticated:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         ingredients = []
-        for el in request.data.get('ingredients'):
+        for el in request.data.get('ingredients', []):
             ingredient = get_object_or_404(Ingredient, id=int(el.get('id')))
             ingredient_for_recipe = IngredientForRecipe(
                 ingredient=ingredient,
@@ -138,32 +152,10 @@ class CreateRecipe(ListAPIView, APIView):
         else:
             return Response(status=status.HTTP_400_BAD_REQUEST, data=serializers.errors)
 
-    def check_filter(self, filter_is_favorite, is_in_shopping_cart, filter_author, filter_tags, recipe, shopping_list,
-                     favorites_list, author, tags):
-        if filter_is_favorite is not None and (
-                filter_is_favorite
-                and recipe.get('id') not in favorites_list
-                or not filter_is_favorite
-                and recipe.get('id') in favorites_list
-        ):
-            return True
-        if is_in_shopping_cart is not None and (
-                is_in_shopping_cart
-                and recipe.get('id') not in shopping_list
-                or not is_in_shopping_cart
-                and recipe.get('id') in shopping_list
-        ):
-            return True
-        if filter_author is not None and filter_author != author.id:
-            return True
-        if filter_tags is not None and filter_tags is not tags:
-            return True
-        return False
-
     def list(self, request, *args, **kwargs):
         data = []
         subscribing_authors, favorites_list, shopping_list = get_lists(request)
-        filter_is_favorite, is_in_shopping_cart, filter_author, filter_tags = self.filters(request=request)
+        filter_is_favorite, is_in_shopping_cart, filter_author, filter_tags = filters(request=request)
         for recipe in self.get_serializer(self.get_queryset(), many=True).data:
             author = CustomUser.objects.get(id=recipe.get('author'))
             tags = [
@@ -174,7 +166,7 @@ class CreateRecipe(ListAPIView, APIView):
                 ingredient_data(ingredient=ingredient)
                 for ingredient in recipe.get('ingredients')
             ]
-            self.check_filter(
+            check_filter(
                 filter_is_favorite=filter_is_favorite, is_in_shopping_cart=is_in_shopping_cart,
                 filter_author=filter_author,
                 filter_tags=filter_tags, recipe=recipe, shopping_list=shopping_list,
@@ -212,6 +204,8 @@ class RecipeGet(APIView):
         return Response(data=data, status=status.HTTP_200_OK)
 
     def patch(self, request, id):
+        if not request.user.is_authenticated:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         recipe = get_object_or_404(Recipe, id=id)
         update_data = request.data
         ingredients = []
@@ -219,6 +213,9 @@ class RecipeGet(APIView):
             ingredient_for_recipe = get_object_or_404(IngredientForRecipe, id=int(el.get('id')))
             if ingredient_for_recipe.amount == float(el.get('amount')):
                 continue
+            if float(el.get('amount')) < 1:
+                return Response(status=status.HTTP_400_BAD_REQUEST,
+                                data={'amount': 'Убедитесь, что это значение больше либо равно 1.'})
             ingredient_for_recipe.amount = float(el.get('amount'))
             ingredient_for_recipe.save()
             ingredients.append(ingredient_for_recipe.id)
@@ -249,6 +246,8 @@ class RecipeGet(APIView):
         return Response(status=status.HTTP_200_OK, data=data)
 
     def delete(self, request, id):
+        if not request.user.is_authenticated:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         get_object_or_404(Recipe, id=id).delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -260,6 +259,8 @@ class RecipeLict(ListAPIView):
 
 
 class ShoppingListCreate(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request, id):
         recipe = get_object_or_404(Recipe, id=id)
         shopping_list = get_object_or_404(ShoppingList, user=request.user)
@@ -279,9 +280,11 @@ class ShoppingListCreate(APIView):
 
 
 class DownloadShoppingList(APIView):
+    permission_classes = [IsAuthenticated]
+
     def get(self, request):
         shopping_list = get_object_or_404(ShoppingList, user=request.user)
-        name = f'media/files/{request.user.username}.pdf'
+        name = f'media/files/{request.user.username}.'
         data = {}
         for recipe in shopping_list.shopping_list.all():
             for ingredient in recipe.ingredients.all():
@@ -292,8 +295,10 @@ class DownloadShoppingList(APIView):
                         'amount': ingredient.amount,
                         'measurement_unit': ingredient.ingredient.measurement_unit
                     }
-        self.get_pdf(data, name)
-        return Response(status=200)
+        self.get_pdf(data, f'{name}pdf')
+        self.get_txt(data, f'{name}txt')
+        self.get_csv(data, f'{name}csv')
+        return Response(status=200, data=name)
 
     def get_pdf(self, data, name):
         pdf = FPDF(orientation='P', unit='mm', format='A4')
@@ -306,8 +311,24 @@ class DownloadShoppingList(APIView):
             pdf.cell(0, 10, txt=str_file, ln=1, align="L")
         pdf.output(name)
 
+    def get_txt(self, data, name):
+        with open(name, 'w', encoding='utf-8') as f:
+            for el, value in data.items():
+                f.write(f"{el} ({value['measurement_unit']}) - ({data[el]['amount']})\n")
+
+    def get_csv(self, data, name):
+        with open(name, 'w', encoding='utf-8') as f:
+            string_file = ''.join(
+                f"{el} ({value['measurement_unit']}) - ({data[el]['amount']}),"
+                for el, value in data.items()
+            )
+            string_file = string_file[:-1]
+            f.write(string_file)
+
 
 class FavoriteCrud(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request, id):
         recipe = get_object_or_404(Recipe, id=id)
         favorites_list = get_object_or_404(FavoritesList, user=request.user)
@@ -327,6 +348,7 @@ class FavoriteCrud(APIView):
 
 
 class SubscribingAuthorsCrud(APIView):
+
     def get(self, request):
         subscribing_author = get_object_or_404(SubscribingAuthors, user=request.user)
         data = []
@@ -350,6 +372,8 @@ class SubscribingAuthorsCrud(APIView):
 
 
 class SubscribingAuthorsCreate(APIView):
+    permission_classes = [IsAuthenticated]
+
     def post(self, request, id):
         user = get_object_or_404(CustomUser, id=id)
         subscribing_authors = get_object_or_404(SubscribingAuthors, user=request.user)
